Start to clean up websockets
