+++ oada-core/http-handler/oada-srvc-http-handler/config.js
@@ -1,3 +1,3 @@
 module.exports = require('../../libs/oada-lib-config')(
-    require('./config.defaults.js')
+    require('./config.defaults')
 )
