+++ oada-core/http-handler/oada-srvc-http-handler/config.ts
@@ -1,0 +1,5 @@
+import libConfig from '../../libs/oada-lib-config'
+// @ts-ignore
+import defaults from './config.defaults'
+
+export default libConfig(defaults)
