+++ oada-core/http-handler/oada-srvc-http-handler/package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "test": "mocha",
-    "start": "CONTAINER_NAME=\"HTTP Handler\" node server.js"
+    "start": "CONTAINER_NAME=\"HTTP Handler\" ts-node --files server.js"
   },
   "mocha": {
     "require": [
@@ -44,15 +44,22 @@
     "oada-error": "^1.1.1",
     "stream": "^0.0.2",
     "timed-cache": "^1.1.0",
+    "ts-node": "^8.8.1",
     "type-is": "^1.6.18",
     "uuid": "^3.0.1",
     "well-known-json": "^0.2.0"
   },
   "devDependencies": {
+    "@types/debug": "^4.1.5",
+    "@types/jsonpointer": "^4.0.0",
+    "@types/node": "^12.12.31",
+    "@types/ws": "^7.2.3",
     "chai": "^3.5.0",
     "mocha": "^7.0.1",
     "mocha-steps": "^1.0.2",
     "mock-require": "^2.0.2",
-    "supertest": "^3.0.0"
+    "prettier-standard": "^16.2.1",
+    "supertest": "^3.0.0",
+    "typescript": "^3.8.3"
   }
 }
