+++ oada-core/http-handler/oada-srvc-http-handler/resources.js
@@ -2,7 +2,6 @@
 
 global.Promise = require('bluebird')
 const uuid = require('uuid')
-const axios = require('axios')
 const express = require('express')
 const bodyParser = require('body-parser')
 const pointer = require('json-pointer')
@@ -28,7 +27,6 @@ const CACHE_PATH = config.get('storage:binary:cacache')
 var requester = require('./requester')
 
 var router = express.Router()
-var expressWs = require('express-ws')(router)
 
 // Turn POSTs into PUTs at random id
 router.post('/*?', function postResource (req, res, next) {
