+++ oada-core/http-handler/oada-srvc-http-handler/types.d.ts
@@ -1,0 +1,11 @@
+declare module 'oada-error' {
+    export class OADAError {
+        constructor (
+            message: string,
+            code: number,
+            userMessage?: string,
+            href?: string,
+            detail?: string
+        )
+    }
+}
