+++ oada-core/libs/oada-lib-config/index.d.ts
@@ -1,0 +1,3 @@
+import type { Provider } from 'nconf'
+
+export default function config (defaults: any): Provider
