+++ oada-core/libs/oada-lib-kafka/index.d.ts
@@ -1,0 +1,19 @@
+// prettier-ignore
+import type { EventEmitter } from 'events'
+
+declare class Base extends EventEmitter {
+}
+
+interface KafkaReguest {
+    msgtype: string
+    code: string
+}
+
+export class Responder extends Base {
+    constructor (
+        consumeTopic: string,
+        produceTopic: string | null,
+        group: string
+    )
+    on(event: 'request', cb: (reg: KafkaReguest) => any): this;
+}
